+++ Assets/Scenes/Main VR.unity
@@ -1193,7 +1193,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 783743726}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 887c49cd57edafa4181892f2ea222103, type: 3}
+  m_Script: {fileID: 11500000, guid: 98b6cee45036b81408ab08addfcc8ac5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   pinchAnimationAction:
@@ -1202,7 +1202,7 @@ MonoBehaviour:
       m_Name: Pinch Animation
       m_Type: 0
       m_ExpectedControlType: 
-      m_Id: 9bf420d5-5d2a-40f2-a3ef-42d3dc26b216
+      m_Id: da1332c1-3403-4783-9baf-84e1f8542063
       m_Processors: 
       m_Interactions: 
       m_SingletonActionBindings: []
@@ -1210,18 +1210,17 @@ MonoBehaviour:
     m_Reference: {fileID: 7904272356298805229, guid: c348712bda248c246b8c49b3db54643f,
       type: 3}
   gripAnimationAction:
-    m_UseReference: 1
+    m_UseReference: 0
     m_Action:
       m_Name: Grip Animation
       m_Type: 0
       m_ExpectedControlType: 
-      m_Id: 2ad57f32-92e2-48e8-a0e5-de991c26c7f0
+      m_Id: fd3632f4-9861-4d11-82b6-13595b42a0fd
       m_Processors: 
       m_Interactions: 
       m_SingletonActionBindings: []
       m_Flags: 0
-    m_Reference: {fileID: -1758520528963094988, guid: c348712bda248c246b8c49b3db54643f,
-      type: 3}
+    m_Reference: {fileID: 0}
   handAnimator: {fileID: 783743728}
 --- !u!95 &783743728 stripped
 Animator:
@@ -2371,6 +2370,11 @@ PrefabInstance:
       propertyPath: m_Name
       value: Left Hand Model
       objectReference: {fileID: 0}
+    - target: {fileID: 109434732395284509, guid: 65847589f926b734d9eaed6e25f0046b,
+        type: 3}
+      propertyPath: m_TagString
+      value: Hand
+      objectReference: {fileID: 0}
     - target: {fileID: 109434732395451517, guid: 65847589f926b734d9eaed6e25f0046b,
         type: 3}
       propertyPath: m_RootOrder
@@ -2443,7 +2447,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1030171710}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 887c49cd57edafa4181892f2ea222103, type: 3}
+  m_Script: {fileID: 11500000, guid: 98b6cee45036b81408ab08addfcc8ac5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   pinchAnimationAction:
@@ -2452,7 +2456,7 @@ MonoBehaviour:
       m_Name: Pinch Animation
       m_Type: 0
       m_ExpectedControlType: 
-      m_Id: c9ad4e17-dc8c-4adf-9f52-54c54e12977e
+      m_Id: 0f5a5194-ff79-4358-b2ba-19d0211e3abf
       m_Processors: 
       m_Interactions: 
       m_SingletonActionBindings: []
@@ -2460,18 +2464,17 @@ MonoBehaviour:
     m_Reference: {fileID: -4289430672226363583, guid: c348712bda248c246b8c49b3db54643f,
       type: 3}
   gripAnimationAction:
-    m_UseReference: 1
+    m_UseReference: 0
     m_Action:
       m_Name: Grip Animation
       m_Type: 0
       m_ExpectedControlType: 
-      m_Id: bdf6519c-3ee3-4915-aac4-d7475a17de40
+      m_Id: e424ee43-2841-459e-8161-6f2e200ec32b
       m_Processors: 
       m_Interactions: 
       m_SingletonActionBindings: []
       m_Flags: 0
-    m_Reference: {fileID: 6558622148059887818, guid: c348712bda248c246b8c49b3db54643f,
-      type: 3}
+    m_Reference: {fileID: 0}
   handAnimator: {fileID: 1983170930}
 --- !u!95 &1983170930 stripped
 Animator:
@@ -2555,6 +2558,16 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: -2899742431399397401, guid: f99572b5d92480447a2097dea2a74740,
+        type: 3}
+      propertyPath: pinchAnimationAction.m_Action.m_Id
+      value: f3fac202-2c16-47d0-890a-373ded606c95
+      objectReference: {fileID: 0}
+    - target: {fileID: -2899742431399397401, guid: f99572b5d92480447a2097dea2a74740,
+        type: 3}
+      propertyPath: pinchAnimationAction.m_Action.m_Name
+      value: Pinch Animation
+      objectReference: {fileID: 0}
     - target: {fileID: 5723723860043569128, guid: f99572b5d92480447a2097dea2a74740,
         type: 3}
       propertyPath: m_InteractionManager
