+++ Assets/Scripts/AnimateHandOnInput.cs
@@ -5,23 +5,26 @@ using UnityEngine.InputSystem;
 
 public class AnimateHandOnInput : MonoBehaviour
 {
-    public InputActionProperty pinchAnimationAction;
-    public InputActionProperty gripAnimationAction;
-    public Animator handAnimator;
+  public InputActionProperty pinchAnimationAction;
+  public InputActionProperty gripAnimationAction;
+  public Animator handAnimator;
 
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
 
-    // Update is called once per frame
-    void Update()
-    {
-        float triggerValue = pinchAnimationAction.action.ReadValue<float>();
-        handAnimator.SetFloat("Trigger", triggerValue);
+  // Start is called before the first frame update
+  void Start()
+  {
 
-        float gripValue = gripAnimationAction.action.ReadValue<float>();
-        handAnimator.SetFloat("Grip", gripValue);
-    }
+  }
+
+  // Update is called once per frame
+  void Update()
+  {
+    float triggerValue = pinchAnimationAction.action.ReadValue<float>();
+    handAnimator.SetFloat("Trigger", triggerValue);
+
+    float gripValue = gripAnimationAction.action.ReadValue<float>();
+    handAnimator.SetFloat("Grip", gripValue);
+
+  }
+
 }
